# backend/openapi.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "Dayonme API",
      "description": "Dayonme 서비스의 API 명세",
      "version": "1.0.0"
    },
    "servers": [
      {
        "url": "/api",
        "description": "API 기본 경로"
      }
    ],
    "components": {
      "securitySchemes": {
        "bearerAuth": {
          "type": "http",
          "scheme": "bearer",
          "bearerFormat": "JWT"
        }
      },
      "schemas": {
        "Error": {
          "type": "object",
          "properties": {
            "status": {
              "type": "string",
              "example": "error"
            },
            "message": {
              "type": "string"
            }
          }
        },
...
        "Emotion": {
          "type": "object",
          "properties": {
            "emotion_id": {
              "type": "integer"
            },
            "name": {
              "type": "string"
            },
            "icon": {
              "type": "string"
            },
            "color": {
              "type": "string"
            }
          }
        },
        "Post": {
          "type": "object",
          "properties": {
            "post_id": {
              "type": "integer"
            },
            "user_id": {
              "type": "integer"
            },
            "content": {
              "type": "string"
            },
            "emotion_summary": {
              "type": "string",
              "nullable": true
            },
            "image_url": {
              "type": "string",
              "nullable": true
            },
            "is_anonymous": {
              "type": "boolean"
            },
            "like_count": {
              "type": "integer"
            },
            "comment_count": {
              "type": "integer"
            },
            "emotions": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/Emotion"
              }
            },
            "created_at": {
              "type": "string",
              "format": "date-time"
            },
            "updated_at": {
              "type": "string",
              "format": "date-time"
            }
          }
        },
...
        "Pagination": {
          "type": "object",
          "properties": {
            "current_page": {
              "type": "integer"
            },
            "total_pages": {
              "type": "integer"
            },
            "total_count": {
              "type": "integer"
            },
            "has_next": {
              "type": "boolean"
            }
          }
        }
      }
    },
    "security": [
      {
        "bearerAuth": []
      }
    ],
    "paths": {
...
      "/my-day/posts": {
        "post": {
          "tags": ["MyDay"],
          "summary": "나의 하루 게시물 생성",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["content"],
                  "properties": {
                    "content": {
                      "type": "string",
                      "minLength": 10,
                      "maxLength": 1000
                    },
                    "emotion_summary": {
                      "type": "string",
                      "maxLength": 100
                    },
                    "image_url": {
                      "type": "string"
                    },
                    "is_anonymous": {
                      "type": "boolean"
                    },
                    "emotion_ids": {
                      "type": "array",
                      "items": {
                        "type": "integer"
                      }
                    }
                  }
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "게시물 생성 성공",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "status": {
                        "type": "string",
                        "example": "success"
                      },
                      "message": {
                        "type": "string",
                        "example": "작업이 성공적으로 완료되었습니다."
                      },
                      "data": {
                        "type": "object",
                        "properties": {
                          "post_id": {
                            "type": "integer"
                          }
                        }
                      }
                    }
                  }
                }
              }
            },
            "400": {
              "description": "잘못된 요청",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Error"
                  }
                }
              }
            },
            "401": {
              "description": "인증 필요",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Error"
                  }
                }
              }
            }
          }
        },
        "get": {
          "tags": ["MyDay"],
          "summary": "나의 하루 게시물 목록 조회",
          "parameters": [
            {
              "name": "page",
              "in": "query",
              "schema": {
                "type": "integer",
                "default": 1
              }
            },
            {
              "name": "limit",
              "in": "query",
              "schema": {
                "type": "integer",
                "default": 10
              }
            },
            {
              "name": "sort_by",
              "in": "query",
              "schema": {
                "type": "string",
                "enum": ["latest", "popular"],
                "default": "latest"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "게시물 목록 조회 성공",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "status": {
                        "type": "string",
                        "example": "success"
                      },
                      "data": {
                        "type": "object",
                        "properties": {
                          "posts": {
                            "type": "array",
                            "items": {
                              "$ref": "#/components/schemas/Post"
                            }
                          },
                          "pagination": {
                            "$ref": "#/components/schemas/Pagination"
                          }
                        }
                      }
                    }
                  }
                }
              }
            },
            "401": {
              "description": "인증 필요",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Error"
                  }
                }
              }
            }
          }
        }
      },
      "/my-day/posts/{id}": {
        "get": {
          "tags": ["MyDay"],
          "summary": "특정 나의 하루 게시물 조회",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "게시물 조회 성공",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "status": {
                        "type": "string",
                        "example": "success"
                      },
                      "data": {
                        "$ref": "#/components/schemas/Post"
                      }
                    }
                  }
                }
              }
            },
            "404": {
              "description": "게시물을 찾을 수 없음",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Error"
                  }
                }
              }
            }
          }
        },
        "delete": {
          "tags": ["MyDay"],
          "summary": "나의 하루 게시물 삭제",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "게시물 삭제 성공",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "status": {
                        "type": "string",
                        "example": "success"
                      },
                      "message": {
                        "type": "string",
                        "example": "게시물이 삭제되었습니다."
                      }
                    }
                  }
                }
              }
            },
            "401": {
              "description": "인증 필요",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Error"
                  }
                }
              }
            },
            "403": {
              "description": "권한 없음",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Error"
                  }
                }
              }
            },
            "404": {
              "description": "게시물을 찾을 수 없음",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Error"
                  }
                }
              }
            }
          }
        }
      },
...
      }
    }